fix: Use standard %2F encoding for autocomplete history and see-all links
Replaces legacy $2F encoding with standard percent-encoding (%2F) for
both history item links and the "see all" link in autocomplete dropdown,
matching the search bar's encoding behavior since vtex.store-components@0e9a0ee5.

Changes:
- Create react/utils/term-encoding.ts with buildHistoryItemValue and
  buildSearchTermRouteValue helpers (both backed by safeEncodeURIComponent)
- Update Autocomplete#updateHistory to use buildHistoryItemValue for
  history item values
- Update Autocomplete#contentWhenQueryIsNotEmpty to use
  buildSearchTermRouteValue for the "see all" link
- Remove react/utils/history-items.ts (functionality moved to term-encoding.ts)
- Add comprehensive test coverage in term-encoding.test.ts
- Mark encodeUrlString as deprecated (no live callers after this fix)
- Update CHANGELOG.md

Spec: is-io-specs/specs/fix-autocomplete-history-link-encoding/spec.md
